feat(remap): Enfore consistant remapping
We want to be able to bake out the remapping, and then reprocess
enforcing the original remap. To be able to do this we need mappings
to be 1 to 1 (so that we do not have to guess which possible mapping
applied for any one image).

This also splits out templating of the new image names and the actual
Copy of images into separate remappers (we may split the copying out
entirely).

Signed-off-by: Tristan Colgate-McFarlane <tristan@cerbos.dev>